Added tests for http nukleus acting as a client.
