+++ src/test/java/org/reaktivity/nukleus/http/internal/streams/rfc7230/MessageFormatLimitsIT.java
@@ -13,27 +13,35 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.nukleus.http.internal.streams.server.rfc7230;
+package org.reaktivity.nukleus.http.internal.streams.rfc7230;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.rules.RuleChain.outerRule;
+import static org.reaktivity.nukleus.http.internal.Context.MAXIMUM_HEADERS_SIZE_PROPERTY_NAME;
+import static org.reaktivity.nukleus.http.internal.Context.MEMORY_FOR_DECODE_PROPERTY_NAME;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.DisableOnDebug;
 import org.junit.rules.TestRule;
 import org.junit.rules.Timeout;
+import org.kaazing.k3po.junit.annotation.ScriptProperty;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
+import org.reaktivity.nukleus.http.internal.test.SystemPropertiesRule;
 import org.reaktivity.reaktor.test.NukleusRule;
 
-public class ArchitectureIT
+public class MessageFormatLimitsIT
 {
     private final K3poRule k3po = new K3poRule()
             .addScriptRoot("route", "org/reaktivity/specification/nukleus/http/control/route")
-            .addScriptRoot("streams", "org/reaktivity/specification/nukleus/http/streams/rfc7230/architecture");
+            .addScriptRoot("streams", "org/reaktivity/specification/nukleus/http/streams/rfc7230/message.format");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
+
+    private final TestRule properties = new SystemPropertiesRule()
+        .setProperty(MAXIMUM_HEADERS_SIZE_PROPERTY_NAME, "64")
+        .setProperty(MEMORY_FOR_DECODE_PROPERTY_NAME, "64");
 
     private final NukleusRule nukleus = new NukleusRule("http")
         .directory("target/nukleus-itests")
@@ -47,14 +55,13 @@ public class ArchitectureIT
         .streams("source", "http#target");
 
     @Rule
-    public final TestRule chain = outerRule(nukleus).around(k3po).around(timeout);
+    public final TestRule chain = outerRule(properties).around(nukleus).around(k3po).around(timeout);
 
     @Test
     @Specification({
         "${route}/input/new/controller",
-        "${streams}/request.and.response/server/source",
-        "${streams}/request.and.response/server/target" })
-    public void shouldCorrelateRequestAndResponse() throws Exception
+        "${streams}/request.headers.too.long/server/source" })
+    public void shouldRejectRequestExceedingMaximumHeadersSize() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_INPUT");
@@ -64,9 +71,24 @@ public class ArchitectureIT
 
     @Test
     @Specification({
+        "${route}/output/new/controller",
+        "${streams}/response.headers.too.long/client/source",
+        "${streams}/response.headers.too.long/client/target"})
+    public void shouldRejectResponseExceedingMaximumHeadersSize() throws Exception
+    {
+        k3po.start();
+        k3po.notifyBarrier("ROUTED_OUTPUT");
+        k3po.notifyBarrier("ROUTED_INPUT");
+        k3po.finish();
+    }
+
+    @Test
+    @Specification({
         "${route}/input/new/controller",
-        "${streams}/request.header.host.missing/server/source" })
-    public void shouldRejectRequestWhenHostHeaderMissing() throws Exception
+        "${streams}/request.fragmented.with.content.length/server/source",
+        "${streams}/request.fragmented.with.content.length/server/target" })
+    @ScriptProperty("targetInputInitialWindow [0x40 0x00 0x00 0x00]") // 64 bytes, same as max headers size
+    public void shouldAcceptFragmentedRequestWithDataWhenOnlyDataExceedsMaxHttpHeadersSize() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_INPUT");
@@ -76,63 +98,17 @@ public class ArchitectureIT
 
     @Test
     @Specification({
-        "${route}/input/new/controller",
-        "${streams}/request.version.http.1.2+/server/source",
-        "${streams}/request.version.http.1.2+/server/target" })
-    public void shouldRespondVersionHttp11WhenRequestVersionHttp12plus() throws Exception
+        "${route}/output/new/controller",
+        "${streams}/response.fragmented.with.content.length/client/source",
+        "${streams}/response.fragmented.with.content.length/client/target" })
+    @ScriptProperty("targetInputInitialWindow [0x40 0x00 0x00 0x00]") // 64 bytes, same as max headers size
+    public void shouldAcceptFragmentedResponseWithDataWhenOnlyDataExceedsMaxHttpHeadersSize() throws Exception
     {
         k3po.start();
-        k3po.awaitBarrier("ROUTED_INPUT");
+        k3po.notifyBarrier("ROUTED_INPUT");
         k3po.notifyBarrier("ROUTED_OUTPUT");
         k3po.finish();
     }
 
-    @Test
-    @Specification({
-        "${route}/input/new/controller",
-        "${streams}/request.version.invalid/server/source" })
-    public void shouldRejectRequestWhenVersionInvalid() throws Exception
-    {
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${route}/input/new/controller",
-        "${streams}/request.version.not.http.1.x/server/source" })
-    public void shouldRejectRequestWhenVersionNotHttp1x() throws Exception
-    {
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${route}/input/new/controller",
-        "${streams}/request.uri.with.user.info/server/source", })
-    public void shouldRejectRequestWithUserInfo() throws Exception
-    {
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${route}/input/new/controller",
-        "${streams}/request.uri.with.percent.chars/server/source",
-        "${streams}/request.uri.with.percent.chars/server/target" })
-    public void shouldAcceptRequestWithPercentChars() throws Exception
-    {
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_INPUT");
-        k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
 }
+
